+++ config/settings.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Outerweb/FilamentSettings
-return [
-
-];