+++ resources/lang/en/settings.php
@@ -1,6 +1,21 @@
 <?php
 
-// translations for Outerweb/FilamentSettings
+declare(strict_types=1);
+
 return [
-    //
+    'page' => [
+        'title' => 'Settings',
+    ],
+    'form' => [
+        'actions' => [
+            'save' => [
+                'label' => 'Save changes',
+            ],
+        ],
+    ],
+    'notifications' => [
+        'saved' => [
+            'title' => 'Settings saved successfully.',
+        ],
+    ],
 ];