+++ src/FilamentSettings.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Outerweb\FilamentSettings;
-
-class FilamentSettings {}