+++ tests/TestCase.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Outerweb\FilamentSettings\Tests;
 
 use BladeUI\Heroicons\BladeHeroiconsServiceProvider;
@@ -9,14 +11,17 @@ use Filament\FilamentServiceProvider;
 use Filament\Forms\FormsServiceProvider;
 use Filament\Infolists\InfolistsServiceProvider;
 use Filament\Notifications\NotificationsServiceProvider;
+use Filament\Schemas\SchemasServiceProvider;
 use Filament\Support\SupportServiceProvider;
 use Filament\Tables\TablesServiceProvider;
 use Filament\Widgets\WidgetsServiceProvider;
-use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Facades\Artisan;
 use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
-use RyanChandler\BladeCaptureDirective\BladeCaptureDirectiveServiceProvider;
 use Outerweb\FilamentSettings\FilamentSettingsServiceProvider;
+use Outerweb\FilamentSettings\Tests\Fixtures\Providers\Filament\AdminPanelProvider;
+use Outerweb\Settings\SettingsServiceProvider;
+use RyanChandler\BladeCaptureDirective\BladeCaptureDirectiveServiceProvider;
 
 class TestCase extends Orchestra
 {
@@ -24,14 +29,20 @@ class TestCase extends Orchestra
     {
         parent::setUp();
 
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Outerweb\\FilamentSettings\\Database\\Factories\\' . class_basename($modelName) . 'Factory'
-        );
+        Artisan::call('vendor:publish --tag=settings-migrations');
+        Artisan::call('migrate');
+    }
+
+    public function getEnvironmentSetUp($app)
+    {
+        config()->set('database.default', 'testing');
+        config()->set('app.key', 'base64:'.base64_encode(random_bytes(32)));
     }
 
     protected function getPackageProviders($app)
     {
-        return [
+        return collect([
+            AdminPanelProvider::class,
             ActionsServiceProvider::class,
             BladeCaptureDirectiveServiceProvider::class,
             BladeHeroiconsServiceProvider::class,
@@ -41,20 +52,19 @@ class TestCase extends Orchestra
             InfolistsServiceProvider::class,
             LivewireServiceProvider::class,
             NotificationsServiceProvider::class,
+            SchemasServiceProvider::class,
             SupportServiceProvider::class,
             TablesServiceProvider::class,
             WidgetsServiceProvider::class,
+            SettingsServiceProvider::class,
             FilamentSettingsServiceProvider::class,
-        ];
+        ])
+            ->sort()
+            ->toArray();
     }
 
-    public function getEnvironmentSetUp($app)
+    protected function defineDatabaseMigrations(): void
     {
-        config()->set('database.default', 'testing');
-
-        /*
-        $migration = include __DIR__.'/../database/migrations/create_filament-settings_table.php.stub';
-        $migration->up();
-        */
+        $this->loadLaravelMigrations();
     }
 }